core-1555 refactor analysis info and param service calls.
