+++ src/components/analyses/details/ParamsPanel.js
@@ -6,38 +6,7 @@
  *
  */
 
-import React, { useState } from "react";
-
-import Link from "next/link";
-import { useQuery } from "react-query";
-
-import { useTranslation } from "i18n";
-
-import ids from "../ids";
-import { isInputType, isReferenceGenomeType } from "./ArgumentTypeUtils";
-
 import {
-    getFolderPage,
-    getParentPath,
-    parseNameFromPath,
-} from "components/data/utils";
-import AppParamTypes from "components/models/AppParamTypes";
-
-import DELink from "components/utils/DELink";
-import DETableHead from "components/table/DETableHead";
-import ErrorTypographyWithDialog from "components/error/ErrorTypographyWithDialog";
-import { ERROR_CODES, getErrorCode } from "components/error/errorCode";
-import TableLoading from "components/table/TableLoading";
-
-import {
-    getResourceDetails,
-    DATA_DETAILS_QUERY_KEY,
-} from "serviceFacades/filesystem";
-
-import buildID from "components/utils/DebugIDUtil";
-
-import {
-    CircularProgress,
     Table,
     TableBody,
     TableCell,
@@ -45,75 +14,15 @@ import {
     TableRow,
     Typography,
 } from "@material-ui/core";
-
-function InputParameterValue(props) {
-    const { id, param_type, path } = props;
-
-    const [statError, setStatError] = React.useState();
-    const [otherError, setOtherError] = React.useState();
-
-    React.useEffect(() => {
-        if (!path) {
-            setStatError(null);
-            setOtherError(null);
-        }
-    }, [path]);
-
-    const displayValue = parseNameFromPath(path);
-    const linkTarget = getFolderPage(
-        AppParamTypes.FOLDER_INPUT === param_type ? path : getParentPath(path)
-    );
-
-    const { t } = useTranslation("analyses");
-
-    const { isFetching: isFetchingStat } = useQuery({
-        queryKey: [DATA_DETAILS_QUERY_KEY, { paths: [path] }],
-        queryFn: () => getResourceDetails({ paths: [path] }),
-        enabled: !!path,
-        onSuccess: () => {
-            setStatError(null);
-            setOtherError(null);
-        },
-        onError: (error) => {
-            if (ERROR_CODES.ERR_DOES_NOT_EXIST === getErrorCode(error)) {
-                setStatError(t("errorInputDoesNotExist", { path: linkTarget }));
-                setOtherError(null);
-            } else {
-                setStatError(null);
-                setOtherError(error);
-            }
-        },
-    });
-
-    return (
-        <>
-            <Typography component="div">
-                {isFetchingStat && <CircularProgress size={16} />}
-                {isFetchingStat || statError ? (
-                    displayValue
-                ) : (
-                    <Link passHref href={linkTarget}>
-                        <DELink
-                            id={id}
-                            text={displayValue}
-                            title={linkTarget}
-                        />
-                    </Link>
-                )}
-            </Typography>
-            <Typography variant="caption" color="error">
-                {statError}
-            </Typography>
-            {otherError && (
-                <ErrorTypographyWithDialog
-                    baseId={id}
-                    errorMessage={t("errorCheckingInputStat")}
-                    errorObject={otherError}
-                />
-            )}
-        </>
-    );
-}
+import ErrorTypographyWithDialog from "components/error/ErrorTypographyWithDialog";
+import DETableHead from "components/table/DETableHead";
+import TableLoading from "components/table/TableLoading";
+import buildID from "components/utils/DebugIDUtil";
+import { useTranslation } from "i18n";
+import React, { useState } from "react";
+import ids from "../ids";
+import { isInputType, isReferenceGenomeType } from "./ArgumentTypeUtils";
+import DataPathLink from "./DataPathLink";
 
 function ParameterValue(props) {
     const {
@@ -130,7 +39,7 @@ function ParameterValue(props) {
 
     if (isInputType(param_type) && valid_info_type) {
         return (
-            <InputParameterValue
+            <DataPathLink
                 id={buildID(baseId, param_id)}
                 param_type={param_type}
                 path={displayValue}
